requirements.txt
torch
pillow
transformers
pyexiftool
pytesseract
ollama
